dbt-airflow-lib/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlh-airflow-common"
dynamic = ["version"]
description = "Common Airflow operators and utilities library"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "MIT"
authors = [
    {name = "Your Team", email = "team@example.com"}
]
keywords = ["airflow", "operators", "workflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "apache-airflow>=3.1.0,<4.0.0",
    "PyYAML>=6.0",
    "dbt-core>=1.8.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "tox>=4.0.0",
    "build>=1.0.0",
    "twine>=5.0.0",
    "setuptools-scm>=8.0.0",
]
dremio = ["dbt-dremio>=1.8.0"]
spark = ["dbt-spark>=1.8.0"]
integration = [
    "dbt-duckdb>=1.8.0",
    "dag-factory>=0.18.0",
]

[project.scripts]
validate-dags = "dlh_airflow_common.validation.cli:main"

[project.urls]
Homepage = "https://github.com/yourorg/dlh-airflow-common"
Repository = "https://github.com/yourorg/dlh-airflow-common"

[tool.setuptools.packages.find]
where = ["src"]

# Setuptools-scm configuration
[tool.setuptools_scm]
write_to = "src/dlh_airflow_common/_version.py"
local_scheme = "no-local-version"
fallback_version = "0.0.0"
search_parent_directories = true


[tool.setuptools.package-data]
dlh_airflow_common = ["py.typed"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
    "src/dlh_airflow_common/_version.py",  # Generated by setuptools-scm
]
# extend-select = ["I"] # isort
# extend-ignore = ["E203", "E501"] # Black compatibility

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Ignore unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["dlh_airflow_common"]

# Mypy configuration
[tool.mypy]
python_version = "3.11"  # Minimum version actively tested (py311, py312)
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "airflow.*"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-m", "not integration",  # Skip integration tests by default
    "--cov=dlh_airflow_common",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=90",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
fail_under = 100
show_missing = true
skip_covered = false
exclude_lines = [
    # Standard exclusions
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    # Async/deferrable code exclusions (requires triggerer environment to test)
    "async def run",
    "def _execute_deferrable",
    "def execute_complete",
    "def on_kill",
]
